add create org memeber screen UI

+++ lib/utils/validation.dart
@@ -13,7 +13,7 @@ String? Function(String?) numberField(String fieldName) {
   return (value) {
     if (value != null &&
         value.trim().isNotEmpty &&
-        int.tryParse(value) == null) {
+        double.tryParse(value) == null) {
       return '$fieldName must be a number';
     }
     return null;
